CRUD and exceptions implemented

+++ frello/src/main/java/com/projectman/frello/services/ProjectService.java
@@ -1,6 +1,7 @@
 package com.projectman.frello.services;
 
 import com.projectman.frello.domain.Project;
+import com.projectman.frello.exceptions.ProjectIdException;
 import com.projectman.frello.repositories.ProjectRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,12 +15,22 @@ public class ProjectService {
     private ProjectRepository projectRepository;
 
     public Project addorUpdateProject(Project project){
-        project.setProjectIdentifier(project.getProjectIdentifier().toUpperCase());
-        return projectRepository.save(project);
+        try{
+            project.setProjectIdentifier(project.getProjectIdentifier().toUpperCase());
+            return projectRepository.save(project);
+        }
+        catch (Exception e) {
+            throw new ProjectIdException("Project ID '"+project.getProjectIdentifier().toUpperCase()+"' already exists");
+        }
     }
 
     public Project findProjectByIdentifier(String projectId){
         Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
+
+        if(project == null){
+            throw new ProjectIdException("Project ID '"+projectId+"' does not exist");
+
+        }
 
         return project;
     }
@@ -30,6 +41,10 @@ public class ProjectService {
 
     public void deleteProjectByIdentifier(String projectId){
         Project project = projectRepository.findByProjectIdentifier(projectId.toUpperCase());
+
+        if(project == null){
+            throw  new  ProjectIdException("Cannot Project with ID '"+projectId+"'. This project does not exist");
+        }
 
         projectRepository.delete(project);
     }
